chore: add GPL3 License with EXCEPTIONS
This adds a GPL3 or later License with exceptions that are conducive to
game modding. Specifically, the exceptions:
1. Allow the linking with Starfield and SFSE without requiring the
distribution of their source code or that they be treated as GPL too.
Starfield is proprietary and SFSE is still considering license choices.
2. Allow the linking of other libraries without requiring that those
libraries be distributed as GPL, but the library's source code must be
provided. Please note, those libraries may be non-free but free
alternatives may be possible given the access to the source code.

Otherwise, the terms of GPL-3-or-later apply. See the FSF FAQ at:
https://www.gnu.org/licenses/gpl-faq.en.html