refactor: main CI workflow (#45)
The previous config seemed to use an older version of both MSVC and
clang-cl which caused failures on including new headers like
`<stdfloat>`--this avoids the errors as it uses newer versions of the
toolsets. Also bumped versions for actions. Tested for release builds
with MSVC/Ninja and clang-cl/Ninja