feat: add MenuInterface + enum names (#98)
There are no `IMenu` classes yet so this is just the `SFSE` machinery.

The version number for `MessagingInterface` was incorrectly set to 2, it
should be 1 as currently set in `SFSE`. This was likely a result of copy
& paste from CLibSSE.

PS: I also named the enums for type discovery. I didn't make them `enum
class` as that would be a breaking change and overkill.