feat: Pattern matching functionality (#109)
Couldn't find this in the lib anywhere, so I'm guessing it got lost
somehow in the Starfield transition. This is the same as Ryan's code,
but I moved it out to a separate file for organization purposes. I also
edited the error message to reflect the possibility that two installed
mods may be incompatible with each other.