fix: Updated VendorModule to detect kMsStore platform correctly and fixed the path accordingly. (#135)
While testing the Address Lib with `SFSE` I noticed that CommonLibSF
wasn't detecting the `kMsStore` platform correctly as `XGameRuntime`
does not exist in the game folder. Changing this to `Xcurl` solved the
issue. I tweaked the path to remove an extra '/' after the game version.

Hopefully I followed the contrib guidlines correctly, apologies if I
didnt.

Co-authored-by: Gazzamc <gazzamc@live.com>